imx: sip: Add imx GPC SIP handler
In order to control the IMR IRQ0 for waking up only the target cores,
and since this is part of the GPC, the IMX_SIP_GPC_CORE_WAKE is being
added as part of the IMX_SIP_GPC group. The new imx_gpc_handler, calls
into imx_a53_core_wake (if the actual SIP id is IMX_SIP_GPC_CORE_WAKE)
which then wakes up all the cores specified in the cpu mask.

Signed-off-by: Abel Vesa <abel.vesa@nxp.com>